# swagger.yaml
openapi: 3.0.3
info:
    title: PoleWin API
    version: 1.0.0
    description: >
        API backend PoleWin (F1) — Express 5 + TypeScript + PostgreSQL/Sequelize.

        **Authentification** : Bearer JWT (access token 15 min).
        Obtenez un token via `POST /api/v1/auth/login` ou `POST /api/v1/auth/register`,
        puis rafraîchissez-le via `POST /api/v1/auth/refresh` (refresh token 7 jours).
servers:
    - url: http://localhost:8000
      description: Serveur local
...
paths:
...
    /api/v1/auth/register:
        post:
            tags: [Auth]
            summary: Créer un compte
            description: >
                Crée un compte utilisateur. Le mot de passe doit respecter les règles de sécurité :
                8 caractères minimum, au moins une majuscule, une minuscule, un chiffre et un caractère spécial.
                Retourne un access token (15 min) et un refresh token (7 jours).
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/RegisterRequest"
                        example:
                            email: "alex@example.com"
                            username: "alexf1"
                            password: "Passw0rd!"
                            confirmPassword: "Passw0rd!"
                            country: "France"
                            language: "fr"
            responses:
                "201":
                    description: Compte créé
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/AuthTokensResponse"
                "400":
                    description: Validation échouée ou email/username déjà utilisé
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "429":
                    description: Trop de requêtes (rate limit)
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
    /api/v1/auth/login:
        post:
            tags: [Auth]
            summary: Connexion
            description: >
                Connexion par email ou username. Retourne un access token (15 min)
                et un refresh token (7 jours).
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/LoginRequest"
                        example:
                            identifier: "alex@example.com"
                            password: "Passw0rd!"
            responses:
                "200":
                    description: Connexion réussie
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/AuthTokensResponse"
                "401":
                    description: Identifiants invalides
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "429":
                    description: Trop de requêtes (rate limit)
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
    /api/v1/auth/refresh:
        post:
            tags: [Auth]
            summary: Rafraîchir les tokens
            description: >
                Échange un refresh token valide contre un nouvel access token (15 min)
                et un nouveau refresh token (7 jours). Rotation à chaque appel.
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/RefreshTokenRequest"
            responses:
                "200":
                    description: Tokens rafraîchis
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/RefreshResponse"
                "401":
                    description: Refresh token invalide ou expiré
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
    /api/v1/auth/verify-email:
        get:
            tags: [Auth]
            summary: Vérifier l'adresse email
            description: Valide le token de vérification d'email envoyé à l'inscription.
            parameters:
                - name: token
                  in: query
                  required: true
                  description: Token JWT de vérification d'email
                  schema:
                      type: string
            responses:
                "200":
                    description: Email vérifié avec succès
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/StatusOkResponse"
                "400":
                    description: Token invalide ou expiré
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
    /api/v1/auth/resend-verify-email:
        post:
            tags: [Auth]
            summary: Renvoyer l'email de vérification
            description: Renvoie un token de vérification d'email. Requiert d'être connecté.
            security:
                - bearerAuth: []
            responses:
                "200":
                    description: Email renvoyé
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/StatusOkResponse"
                "400":
                    description: Email déjà vérifié
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "401":
                    description: JWT manquant ou invalide
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
    /api/v1/auth/change-password:
        post:
            tags: [Auth]
            summary: Changer le mot de passe
            description: >
                Modifie le mot de passe de l'utilisateur connecté.
                Le nouveau mot de passe doit respecter les règles de sécurité.
            security:
                - bearerAuth: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/ChangePasswordRequest"
            responses:
                "200":
                    description: Mot de passe modifié
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/StatusOkResponse"
                "400":
                    description: Validation échouée ou mot de passe actuel incorrect
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "401":
                    description: JWT manquant ou invalide
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
    /api/v1/auth/forgot-password:
        post:
            tags: [Auth]
            summary: Demander la réinitialisation du mot de passe
            description: >
                Envoie un email de réinitialisation si l'adresse est connue.
                Retourne toujours 200 pour éviter l'énumération d'emails.
                Le token de réinitialisation expire après 1 heure.
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/ForgotPasswordRequest"
                        example:
                            email: "alex@example.com"
            responses:
                "200":
                    description: Demande traitée (email envoyé si l'adresse existe)
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/StatusOkResponse"
                "429":
                    description: Trop de requêtes (rate limit)
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
    /api/v1/auth/reset-password:
        post:
            tags: [Auth]
            summary: Réinitialiser le mot de passe
            description: >
                Réinitialise le mot de passe à partir du token reçu par email.
                Le token expire après 1 heure. Le nouveau mot de passe doit respecter
                les règles de sécurité.
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/ResetPasswordRequest"
            responses:
                "200":
                    description: Mot de passe réinitialisé avec succès
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/StatusOkResponse"
                "400":
                    description: Token invalide/expiré ou validation échouée
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "429":
                    description: Trop de requêtes (rate limit)
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
...
    /api/v1/users/me:
        get:
            tags: [Utilisateurs]
            summary: Mon profil complet
            description: Retourne le compte courant avec profil et rôles.
            security:
                - bearerAuth: []
            responses:
                "200":
                    description: Profil complet
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/UserMeResponse"
                "401":
                    description: JWT manquant ou invalide
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"

        patch:
            tags: [Utilisateurs]
            summary: Mettre à jour mon profil
            description: Patch partiel du compte et/ou du profil utilisateur.
            security:
                - bearerAuth: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/UpdateMeRequest"
            responses:
                "200":
                    description: Profil mis à jour
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/UserMeResponse"
                "400":
                    description: Validation échouée
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "401":
                    description: JWT manquant ou invalide
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "409":
                    description: Conflit (username déjà utilisé)
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
...
components:
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
            description: "JWT dans l'en-tête Authorization: Bearer <token>"
    schemas:
...
        PasswordStrong:
            type: string
            minLength: 8
            maxLength: 200
            description: >
                Mot de passe fort : 8 caractères minimum, au moins une majuscule,
                une minuscule, un chiffre et un caractère spécial (!@#$%^&*…).
            example: "Passw0rd!"

        # ── Requêtes Auth ─────────────────────────────────────────────────────
        RegisterRequest:
            type: object
            required: [email, username, password, confirmPassword]
            properties:
                email:
                    type: string
                    format: email
                    maxLength: 255
                    example: "alex@example.com"
                username:
                    type: string
                    minLength: 3
                    maxLength: 50
                    example: "alexf1"
                password:
                    $ref: "#/components/schemas/PasswordStrong"
                confirmPassword:
                    type: string
                    description: Doit être identique à `password`
                    example: "Passw0rd!"
                dateOfBirth:
                    type: string
                    format: date
                    example: "2001-06-20"
                country:
                    type: string
                    maxLength: 100
                    example: "France"
                language:
                    type: string
                    maxLength: 20
                    example: "fr"
        LoginRequest:
            type: object
            required: [identifier, password]
            properties:
                identifier:
                    type: string
                    minLength: 3
                    maxLength: 255
                    description: Email ou username
                    example: "alex@example.com"
                password:
                    type: string
                    minLength: 1
                    maxLength: 200
                    example: "Passw0rd!"
        RefreshTokenRequest:
            type: object
            required: [refreshToken]
            properties:
                refreshToken:
                    type: string
                    description: Refresh token JWT (7 jours)
                    example: "eyJhbGciOiJIUzI1NiJ9..."
        ChangePasswordRequest:
            type: object
            required: [currentPassword, newPassword, confirmNewPassword]
            properties:
                currentPassword:
                    type: string
                    minLength: 1
                    maxLength: 200
                    example: "OldPassw0rd!"
                newPassword:
                    $ref: "#/components/schemas/PasswordStrong"
                confirmNewPassword:
                    type: string
                    description: Doit être identique à `newPassword`
                    example: "NewPassw0rd!"
        ForgotPasswordRequest:
            type: object
            required: [email]
            properties:
                email:
                    type: string
                    format: email
                    maxLength: 255
                    example: "alex@example.com"
        ResetPasswordRequest:
            type: object
            required: [token, newPassword, confirmNewPassword]
            properties:
                token:
                    type: string
                    description: Token de réinitialisation reçu par email (valide 1h)
                    example: "eyJhbGciOiJIUzI1NiJ9..."
                newPassword:
                    $ref: "#/components/schemas/PasswordStrong"
                confirmNewPassword:
                    type: string
                    description: Doit être identique à `newPassword`
                    example: "NewPassw0rd!"
...
        AuthUser:
            type: object
            properties:
                id:
                    type: string
                    description: BIGINT renvoyé en string
                    example: "1"
                email:
                    type: string
                    format: email
                    example: "alex@example.com"
                username:
                    type: string
                    example: "alexf1"
                roles:
                    type: array
                    items:
                        type: string
                    example: ["user"]
        AuthTokensResponse:
            type: object
            properties:
                ok:
                    type: boolean
                    example: true
                accessToken:
                    type: string
                    description: JWT d'accès (15 min)
                    example: "eyJhbGciOiJIUzI1NiJ9..."
                refreshToken:
                    type: string
                    description: JWT de rafraîchissement (7 jours)
                    example: "eyJhbGciOiJIUzI1NiJ9..."
                verifyEmailToken:
                    type: string
                    description: Token de vérification email (24h) — uniquement à l'inscription
                    example: "eyJhbGciOiJIUzI1NiJ9..."
                user:
                    $ref: "#/components/schemas/AuthUser"
            required: [ok, accessToken, refreshToken, user]
        RefreshResponse:
            type: object
            properties:
                ok:
                    type: boolean
                    example: true
                accessToken:
                    type: string
                    description: Nouvel access token (15 min)
                    example: "eyJhbGciOiJIUzI1NiJ9..."
                refreshToken:
                    type: string
                    description: Nouveau refresh token (7 jours)
                    example: "eyJhbGciOiJIUzI1NiJ9..."
            required: [ok, accessToken, refreshToken]
        StatusOkResponse:
            type: object
            properties:
                status:
                    type: string
                    example: ok
            required: [status]

        # ── Utilisateurs ──────────────────────────────────────────────────────
        UserProfile:
            type: object
            properties:
                displayName:
                    type: string
                    nullable: true
                    example: "Alexandre"
                avatarUrl:
                    type: string
                    format: uri
                    nullable: true
                    example: "https://cdn.polewin.app/avatars/123.png"
                bio:
                    type: string
                    nullable: true
                    example: "Fan de F1 et de statistiques."
                favoriteTeamCode:
                    type: string
                    nullable: true
                    example: "ferrari"
                favoriteDriverCode:
                    type: string
                    nullable: true
                    example: "leclerc"
                timeZone:
                    type: string
                    nullable: true
                    example: "Europe/Paris"
                isProfilePublic:
                    type: boolean
                    example: true
                showStats:
                    type: boolean
                    example: true
        User:
            type: object
            properties:
                id:
                    type: string
                    description: BIGINT renvoyé en string
                    example: "1"
                email:
                    type: string
                    format: email
                    example: "alex@example.com"
                username:
                    type: string
                    example: "alexf1"
                dateOfBirth:
                    type: string
                    format: date
                    nullable: true
                    example: "2001-06-20"
                country:
                    type: string
                    nullable: true
                    example: "France"
                language:
                    type: string
                    nullable: true
                    example: "fr"
                isEmailVerified:
                    type: boolean
                    example: false
                isActive:
                    type: boolean
                    example: true
                lastLoginAt:
                    type: string
                    format: date-time
                    nullable: true
                createdAt:
                    type: string
                    format: date-time
                updatedAt:
                    type: string
                    format: date-time
                profile:
                    $ref: "#/components/schemas/UserProfile"
                roles:
                    type: array
                    items:
                        type: string
                    example: ["user"]
        UserMeResponse:
            type: object
            properties:
                status:
                    type: string
                    example: ok
                user:
                    $ref: "#/components/schemas/User"
            required: [status, user]
...
        UpdateMeRequest:
            type: object
            properties:
                username:
                    type: string
                    minLength: 3
                    maxLength: 50
                country:
                    type: string
                    nullable: true
                    maxLength: 100
                language:
                    type: string
                    nullable: true
                    maxLength: 20
                dateOfBirth:
                    type: string
                    format: date
                    nullable: true
                profile:
                    type: object
                    properties:
                        displayName:
                            type: string
                            nullable: true
                            maxLength: 100
                        avatarUrl:
                            type: string
                            nullable: true
                            maxLength: 500
                            format: uri
                        bio:
                            type: string
                            nullable: true
                            maxLength: 2000
                        favoriteTeamCode:
                            type: string
                            nullable: true
                            maxLength: 50
                        favoriteDriverCode:
                            type: string
                            nullable: true
                            maxLength: 50
                        timeZone:
                            type: string
                            nullable: true
                            maxLength: 64
                        isProfilePublic:
                            type: boolean
                        showStats:
                            type: boolean

        # ── Admin ─────────────────────────────────────────────────────────────
...
        ErrorResponse:
            type: object
            properties:
                status:
                    type: string
                    example: error
                code:
                    type: string
                    example: VALIDATION_ERROR
                message:
                    type: string
                    example: "Données invalides"
            required: [status, message]

        # ── Health ────────────────────────────────────────────────────────────
...